Allow cast_ptr_alignment error for clippy
Clippy is complaining a lot about many of the casts the library is
doing. This happens because when we define something like `pub struct
AvroWriter` to be the opaque type in the C bindings and then we cast
`avro_rs::Writer` from/into it, the compiler has no idea that these two
pointers actually point to structs of the same type.

Unfortunately we cannot use `pub type AvroWriter = Writer` because that
would be translated by cbindgen in `typedef Writer AvroWriter`, with
Writer being a missing symbol (since it is only defined in the
dependency). Instead what we want is `typedef struct AvroWriter
AvroWriter`, which is exactly what `pub struct AvroWriter` is giving us.